Change : Method for migration of apply charges from M1 to M2 added and corresponding test cases has been added/updated.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@10391 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/accounts/dao/AccountsApplyChargesDAO.java
@@ -416,11 +416,11 @@ public class AccountsApplyChargesDAO extends DAO {
 			feeList = new ArrayList();
 		}
 		AccountApplyChargesMaster accountApplyChargesMaster = new AccountApplyChargesMaster();
-		accountApplyChargesMaster.setFeeId(AccountConstants.MISC_FEES);
+		accountApplyChargesMaster.setFeeId(Short.valueOf(AccountConstants.MISC_FEES));
 		accountApplyChargesMaster.setFeeName("Misc Fees");
 		feeList.add(accountApplyChargesMaster);
 		accountApplyChargesMaster = new AccountApplyChargesMaster();
-		accountApplyChargesMaster.setFeeId(AccountConstants.MISC_PENALTY);
+		accountApplyChargesMaster.setFeeId(Short.valueOf(AccountConstants.MISC_PENALTY));
 		accountApplyChargesMaster.setFeeName("Misc Penalty");
 		feeList.add(accountApplyChargesMaster);
 	}
